requirements.txt
black==19.3b0
coverage==4.5.4
coveralls==1.8.2
flake8==3.7.8
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
